requirements.txt
ipykernel
python-dotenv

fuzzywuzzy
python-Levenshtein

pandas
openpyxl
matplotlib
seaborn
plotly

openai
llama-index
langchain
tiktoken

nbformat
google-auth
google-api-python-client
google-cloud-storage
google-cloud-secret-manager==2.12.4
google-cloud-aiplatform==1.36.4
db-dtypes==1.0.1
google-api-core
google-auth
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-crc32c
google-resumable-media
googleapis-common-protos
pandas==1.3.5
pandas_gbq
proto-plus==1.22.0
protobuf==3.20.*
pyarrow==8.0.0
pyasn1==0.4.8
PyGithub==2.1.1
pytz==2022.2.1
shapely==1.8.5
urllib3==1.26.12
gspread
oauth2client